Extract UdpTransportLayer
