+++ node/src/main/scala/coop/rchain/node/effects/package.scala
@@ -150,53 +150,7 @@ package object effects {
                                        ev1: Log[Task],
                                        ev2: Time[Task],
                                        ev3: Metrics[Task]): TransportLayer[Task] =
-    new TransportLayer[Task] {
-
-      val net = new UnicastNetwork(src)
-      import scala.concurrent.duration._
-
-      def roundTrip(msg: ProtocolMessage,
-                    remote: ProtocolNode,
-                    timeout: Duration): Task[CommErr[ProtocolMessage]] =
-        net.roundTrip[Task](msg, remote, timeout)
-
-      def local: Task[ProtocolNode] = net.local.pure[Task]
-
-      def send(msg: ProtocolMessage, peer: PeerNode): Task[CommErr[Unit]] =
-        Task.delay(net.comm.send(msg.toByteSeq, peer))
-
-      def broadcast(msg: ProtocolMessage, peers: Seq[PeerNode]): Task[Seq[CommErr[Unit]]] =
-        peers.toList.traverse(peer => send(msg, peer)).map(_.toSeq)
-
-      private def handle(dispatch: ProtocolMessage => Task[Option[ProtocolMessage]])
-        : Option[ProtocolMessage] => Task[Unit] = _.fold(().pure[Task]) { pm =>
-        for {
-          ti <- Time[Task].nanoTime
-          r1 <- dispatch(pm)
-          r2 <- r1.fold(().pure[Task]) { response =>
-                 pm.sender.fold(Log[Task].error(s"Sender not available for $pm")) { sender =>
-                   send(response, sender) >>= {
-                     case Left(error) =>
-                       Log[Task].warn(
-                         s"Was unable to send response $response for request: $pm, error: $error")
-                     case _ => ().pure[Task]
-                   }
-                 }
-               }
-          tf <- Time[Task].nanoTime
-          _  <- Metrics[Task].record("network-roundtrip-micros", (tf - ti) / 1000)
-        } yield r2
-      }
-
-      def receive(dispatch: ProtocolMessage => Task[Option[ProtocolMessage]]): Task[Unit] =
-        net
-          .receiver[Task]
-          .flatMap(handle(dispatch))
-          .forever
-          .executeAsync
-          .start
-          .void
-    }
+    new UdpTransportLayer(src)
 
   class JLineConsoleIO(console: ConsoleReader) extends ConsoleIO[Task] {
     def readLine: Task[String] = Task.delay {
